LayoutTests/http/tests/security/script-crossorigin-error-event-information.html

<!DOCTYPE html>
<head>
</head>
<body>
<script src="../../js-test-resources/js-test-pre.js"></script>
<script>
window.jsTestIsAsync = true;
description("The test passes if error event gets unsanitized information about the script error.");

window.addEventListener('error', function() {
    shouldBeEqualToString("event.message.match(/SomeError/)[0]", "SomeError");
    shouldBeEqualToString("event.filename", "http://localhost:8000/security/resources/cors-script.php?fail=true");
    shouldBe("event.lineno", "1");
    shouldBe("event.colno", "52");
    shouldBeEqualToString("event.error.toString()", "SomeError");
    finishJSTest();
});

successfullyParsed = true;
</script>
<script crossorigin="anonymous" src="http://localhost:8000/security/resources/cors-script.php?fail=true"></script>
<script src="../../js-test-resources/js-test-post.js"></script>
